Clarify inspiration source in README
Removed the phrase 'loosely based on' to clarify inspiration from MapMap. I had AI model the program after it, but all I got was a shell for the program (layout). None of the functionality came over, as I was hoping it would - to save time. Therefore, other than the UI layout, there's nothing taken from mapmap to this project, at all. None of the code behind or otherwise.